refactor: :recycle: remove post-message-bus package
We do not need to publish post-message-bus as a separate package. Collapsing it into `meso-js` will decrease the surface area of our releases.
