+++ packages/meso-js/test/validators.test.ts
@@ -1,6 +1,5 @@
 import { TransferStatus } from "@meso-network/types";
-import { MessageKind } from "../src";
-import { PostMessageHandlerFn } from "../src/types";
+import { MessageKind, PostMessageHandlerFn } from "../src/types";
 import {
   validateHandlerFunction,
   validateMessage,
