--- packages/post-message-bus/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@meso-network/post-message-bus",
-  "version": "0.0.71",
-  "description": "Pass strongly typed messages across frames.",
-  "type": "module",
-  "main": "./lib/index.js",
-  "types": "./lib/index.d.ts",
-  "files": [
-    "lib",
-    "package.json",
-    "README.md",
-    "CHANGELOG.md"
-  ],
-  "scripts": {
-    "typecheck": "tsc -b",
-    "build": "vite build",
-    "clean": "rimraf lib",
-    "test": "vitest run --silent",
-    "test:watch": "vitest --watch"
-  },
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/meso-network/meso-js.git",
-    "directory": "packages/meso-js"
-  },
-  "devDependencies": {
-    "jsdom": "^22.1.0"
-  }
-}
