--- packages/post-message-bus/src/index.ts
@@ -1,5 +1,0 @@
-import { version } from "../package.json";
-
-export * from "../../types";
-export * from "./createPostMessageBus";
-export const MESO_JS_VERSION = version;
