--- packages/post-message-bus/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "lib",
-    "rootDir": "./",
-    "types": ["vitest", "vitest/globals"]
-  },
-  "references": [{ "path": "../types" }],
-  "include": ["./**/*.ts", "../types/**/*.ts", "./package.json"]
-}
